Update kaprekarUtils.c
got rid of (int)

+++ c_version/kaprekarUtils.c
@@ -33,7 +33,7 @@ int isKaprekar(int n) {
   long square = n * n;
 
   //compute the number of digits of the square
-  int numDigits = (int) log10(square) + 1;
+  int numDigits =log10(square) + 1;
   long modulus = 1;
   long first, second;
